Drop built-in Muxy Help, point Documentation to muxy.app/docs
Removes the in-app Muxy Help window; documentation is the single source of truth. Help menu keeps the external
  links and now opens muxy.app/docs.
